Re-enable access to matrices from SWIG
This entailed exposing "AbstractMatrix" to the CC2 API.
AbstractMatrix can be thought of as a minimal interface exposing just
"diagnostic" capabilities, or as a mostly-opaque data type.  (In a
language with interfaces, I would have defined a "DiagnosableMatrix"
interface.)

This minimal view of Alpha/Beta can now be fetched from Evaluator or
AbstractIntegrator.

+++ src/matrix/SparseMatrix.cpp
@@ -99,12 +99,12 @@ size_t SparseMatrix::AllocatedEntries() const
     return sum;
 }
 
-void SparseMatrix::ToHostMatrix(double** mat, size_t* rows, size_t* cols) const
+void SparseMatrix::ToHostMatrix(double** mat, int* rows, int* cols) const
 {
     const double nan = std::numeric_limits<double>::quiet_NaN();
     *mat = new double[Rows() * Columns()];
-    *rows = Rows();
-    *cols = Columns();
+    *rows = static_cast<int>(Rows());
+    *cols = static_cast<int>(Columns());
     for (size_t i = 0; i < Rows(); i++) {
         for (size_t j = 0; j < Columns(); j++) {
             (*mat)[i * Columns() + j] = IsAllocated(i, j) ? Get(i, j) : nan;
